luna_assistant_backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "luna-chatbot"
version = "0.1.1"  
description = "Chainlit Chat assistant"
readme = "README.md"
requires-python = ">=3.12,<3.13"
license = { text="MIT"}
authors = [{name= "Jonas Besier", email = "jonas.besier.work@gmail.com"}]

dependencies = [
    "fastapi",
    "pydantic-settings",
    "uvicorn",
    "python-dotenv",
    "langchain==1.0.2",
    "langchain-openai",
    "pypdf",
    "langchain-community",
    "langchain-chroma",
    "langsmith"
]

[project.optional-dependencies]
dev = ["pytest", "ruff", "mypy", "isort"]


[tool.ruff]
line-length = 100

[tool.mypy]
strict = true


[tool.setuptools]
include-package-data = true